+++ src/api/auth.ts
@@ -1,14 +1,15 @@
 import type { AxiosPromise } from "axios";
 import { basicAxios, mainAxios } from "./helpers/myaxios";
-import type { ApiAuthBag, SelfInfo } from "./types";
+import type {
+  ApiAuthBag,
+  LoginRequest,
+  RefreshUserIdRequest,
+  SelfInfo,
+} from "./types";
 
 export const checkCredentialsApi = (): AxiosPromise<void> =>
   mainAxios.get(`/auth/check`);
 
-export interface RefreshUserIdRequest {
-  refreshToken: string;
-  userId: number;
-}
 export const getTokenFromRefreshApi = (
   data: RefreshUserIdRequest
 ): AxiosPromise<ApiAuthBag> => basicAxios.post(`/auth/get_access`, data);
@@ -16,11 +17,6 @@ export const getTokenFromRefreshApi = (
 export const invalidateTokenApi = (
   data: RefreshUserIdRequest
 ): AxiosPromise<void> => mainAxios.delete(`/auth/invalidate`, { data });
-
-export interface LoginRequest {
-  username: string;
-  password: string;
-}
 
 export const loginApi = (data: LoginRequest): AxiosPromise<ApiAuthBag> =>
   basicAxios.post(`/auth/login`, data);
