+++ src/api/courses.ts
@@ -1,16 +1,13 @@
 import type { AxiosPromise } from "axios";
 import type { CourseDto, CourseId } from "models/course";
-import type { EmployeeId } from "models/employee";
-import type { Page } from "models/generic";
 import { mainAxios } from "./helpers/myaxios";
 import { preparePageRequest } from "./helpers/preparers";
-import type { PageRequest } from "./types";
-
-export interface GetCoursesByOwnerIdRequest {
-  name?: string;
-  employeeId: EmployeeId;
-  pagination: PageRequest<keyof CourseDto>;
-}
+import type {
+  Page,
+  GetCoursesByOwnerIdRequest,
+  CreateCourseRequest,
+  EditCourseRequest,
+} from "./types";
 
 export const getCoursesByOwnerApi = ({
   employeeId,
@@ -24,16 +21,9 @@ export const getCoursesByOwnerApi = ({
 export const getCourseByIdApi = (id: CourseId): AxiosPromise<CourseDto> =>
   mainAxios.get(`/courses/${id}`);
 
-export interface CreateCourseRequest {
-  name: string;
-  owner: number;
-}
-
 export const postCourseApi = (
   body: CreateCourseRequest
 ): AxiosPromise<CourseDto> => mainAxios.post(`/courses/`, body);
-
-export type EditCourseRequest = CreateCourseRequest;
 
 export const putCourseApi = (
   id: CourseId,
