+++ src/redux/courses/thunk.ts
@@ -1,7 +1,8 @@
 import { createAxiosAsyncThunk } from "@redux/utils";
-import { getCoursesByOwnerApi, GetCoursesByOwnerIdRequest } from "api/courses";
+import { getCoursesByOwnerApi } from "api/courses";
+import { GetCoursesByOwnerIdRequest } from "api/types";
 
-export const coursesGetSelfPageThunk = createAxiosAsyncThunk(
+export const getSelfCoursesPageThunk = createAxiosAsyncThunk(
   "courses/getSelfPage",
   async (args: GetCoursesByOwnerIdRequest) => {
     const data = (await getCoursesByOwnerApi(args)).data;
